// src/app/main/printouts/olevels/olevel-report-forms/models/optional-pdf-sections.ts
export interface OptionalPdfSections{
    subjectTeacherComments: boolean;
    classTeacherComments: boolean;
    classTeacherSignature: boolean;
    houseTeacherComments: boolean;
    houseTeacherSignature: boolean;
    principalComments: boolean;
    principalSignature: boolean;
    examSlot: boolean;
    attendanceReport: boolean;
    competencyAreas: boolean;
    projects: boolean;
    activitiesAndValues: boolean;
    scoreDescriptor: boolean;
    rawScore: boolean;
	yearSummaryReport: boolean;
	gradeDescriptor: boolean;
}